Removed unused code and call api with useMutation #1135
